requirements.txt
customtkinter>=5.2.0
sqlalchemy>=2.0.0
SpeechRecognition>=3.10.0
Pillow>=10.0.0
